+++ crate/operator/handlers/handle_restore_backup.py
@@ -27,7 +27,10 @@ from typing import List
 import kopf
 
 from crate.operator.config import config
-from crate.operator.constants import SnapshotRestoreType
+from crate.operator.constants import (
+    DEFAULT_BACKUP_STORAGE_PROVIDER,
+    SnapshotRestoreType,
+)
 from crate.operator.handlers.handle_update_cratedb import get_backoff
 from crate.operator.operations import (
     AfterClusterUpdateSubHandler,
@@ -76,6 +79,7 @@ async def restore_backup(
     tables = new.get("tables", [])
     partitions = new.get("partitions", [])
     sections = new.get("sections", [])
+    backup_provider = new.get("backupProvider", DEFAULT_BACKUP_STORAGE_PROVIDER)
 
     scheme = get_crash_scheme(cratedb)
     pod_name = get_crash_pod_name(cratedb, name)
@@ -102,6 +106,7 @@ async def restore_backup(
         repository,
         snapshot,
         restore_type,
+        backup_provider,
         tables,
         partitions,
         sections,
@@ -192,6 +197,7 @@ def register_restore_handlers(
     repository: str,
     snapshot: str,
     restore_type: str,
+    backup_provider: str,
     tables: list,
     partitions: list,
     sections: list,
@@ -207,6 +213,7 @@ def register_restore_handlers(
             repository=repository,
             snapshot=snapshot,
             restore_type=restore_type,
+            backup_provider=backup_provider,
             tables=tables,
             partitions=partitions,
             sections=sections,