+++ crate/operator/main.py
@@ -283,7 +283,7 @@ async def service_backup_schedule_update(
     **_kwargs,
 ):
     """
-    Handles updates to the backup schedule for AWS s3 backups.
+    Handles updates to the backup schedule for AWS S3 / Azure Storage backups.
     """
     await raise_on_namespace_terminating(namespace)
     await update_backup_schedule(namespace, name, diff, logger)
@@ -379,7 +379,7 @@ async def enable_cronjob_after_delay(
     delay_start_time = status.get(DELAY_CRONJOB_START)
     if delay_start_time:
         delay = (
-            datetime.datetime.utcnow()
+            datetime.datetime.now(datetime.timezone.utc)
             - datetime.timedelta(
                 seconds=5 if config.TESTING else config.RE_ENABLING_CRONJOB_DELAY
             )